test/postgres_db/requirements.txt
psycopg2-binary==2.9.9
networkx==3.2.1
